dsc,test: relax subclass-facing sizes iface and be stricter internally
The update_sizes now receives 2 uint64_t so the subclass does not have
to worry about integer overflows.
The update_sizes_of_header will check for anything out of range.

+++ test/dsc/descriptor.cpp
@@ -1364,7 +1364,7 @@ namespace {
         void write_struct_specifics_into(IOBase&) override {
             return; // 0 write
         }
-        void update_sizes(uint8_t& isize, [[maybe_unused]] uint32_t& csize) override {
+        void update_sizes(uint64_t& isize, [[maybe_unused]] uint64_t& csize) override {
             isize = assert_u8(internal_data.size());
         }
