+++ test/dsc/descriptor_set.cpp
@@ -1043,7 +1043,7 @@ namespace {
         Dummy(const struct Descriptor::header_t& hdr, BlockArray& cblkarr) : Descriptor(hdr, cblkarr) {}
         void read_struct_specifics_from(IOBase&) override {}
         void write_struct_specifics_into(IOBase&) override {}
-        void update_sizes([[maybe_unused]] uint8_t& isize, [[maybe_unused]] uint32_t& csize) override { }
+        void update_sizes([[maybe_unused]] uint64_t& isize, [[maybe_unused]] uint64_t& csize) override { }
         static std::unique_ptr<Descriptor> create(const struct Descriptor::header_t& hdr, BlockArray& cblkarr) {
             return std::make_unique<Dummy>(hdr, cblkarr);
         }
@@ -2616,7 +2616,7 @@ namespace {
 
         void read_struct_specifics_from(IOBase& io) override { io.readall(idata); }
         void write_struct_specifics_into(IOBase& io) override { io.writeall(idata); }
-        void update_sizes(uint8_t& isize, [[maybe_unused]] uint32_t& csize) override {
+        void update_sizes(uint64_t& isize, [[maybe_unused]] uint64_t& csize) override {
             isize = assert_u8(idata.size());
         }
 
@@ -2870,7 +2870,7 @@ namespace {
         }
 
         public:
-        void /* internal */ update_sizes(uint8_t& isize, uint32_t& csize) override {
+        void /* internal */ update_sizes(uint64_t& isize, uint64_t& csize) override {
             DescriptorSet::update_sizes(isize, csize);
             isize += 2; // count for app's own cookie
         }
